+++ src/test/groovy/ru/vyarus/dropwizard/guice/AutoScanModeTest.groovy
@@ -1,28 +1,28 @@
 package ru.vyarus.dropwizard.guice
 
 import com.google.common.collect.Sets
+import com.google.inject.Inject
 import com.google.inject.Injector
 import com.google.inject.Key
 import com.google.inject.TypeLiteral
 import io.dropwizard.Configuration
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
-import org.junit.Rule
 import ru.vyarus.dropwizard.guice.module.installer.feature.LifeCycleInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.ManagedInstaller
-import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.ResourceInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.TaskInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.admin.AdminFilterInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.admin.AdminServletInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.eager.EagerSingletonInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.health.HealthCheckInstaller
-import ru.vyarus.dropwizard.guice.module.installer.feature.plugin.PluginInstaller
+import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.ResourceInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.provider.JerseyProviderInstaller
+import ru.vyarus.dropwizard.guice.module.installer.feature.plugin.PluginInstaller
 import ru.vyarus.dropwizard.guice.module.installer.internal.FeaturesHolder
 import ru.vyarus.dropwizard.guice.support.AutoScanApplication
 import ru.vyarus.dropwizard.guice.support.TestConfiguration
 import ru.vyarus.dropwizard.guice.support.feature.*
-import ru.vyarus.dropwizard.guice.test.GuiceyAppRule
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
 
 /**
  * Dummy test.
@@ -30,17 +30,19 @@ import ru.vyarus.dropwizard.guice.test.GuiceyAppRule
  * @author Vyacheslav Rusakov
  * @since 31.08.2014
  */
+@UseGuiceyApp(AutoScanApplication)
 class AutoScanModeTest extends AbstractTest {
 
-    @Rule
-    GuiceyAppRule<TestConfiguration> RULE = new GuiceyAppRule<>(AutoScanApplication, null);
+    @Inject
+    FeaturesHolder holder
+    @Inject
+    Bootstrap bootstrap
+    @Inject
+    Injector injector
 
     def "Check auto scan configuration"() {
 
         when: "application started"
-        Injector injector = RULE.getInjector()
-        FeaturesHolder holder = RULE.getBean(FeaturesHolder.class);
-        Bootstrap bootstrap = RULE.getBean(Bootstrap.class);
 
         then: "environment binding done"
         bootstrap