+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Add ability to customize injector creation (required by some guice third party modules, like governator). Contributed by [Nicholas Pace](https://github.com/segfly)
+* Add spock extensions to use injections directly in specification (like spock-guice do)
 
 ### 2.0.0 (2014-11-25)
 * Dropwizard 0.8 integration (as result, no more depends on jersey-guice, but depends on guice-bridge(hk2)).